+++ CHANGES.rst
@@ -16,6 +16,9 @@ New Features
   etc. For example, sample weights. This is achieved by passing the scorers and
   their arguments to :meth:`DataOp.skb.with_scoring`. :pr:`1995` by
   :user:`Jérôme Dockès <jeromedockes>`.
+- The :class:`DropSimilar` transformer has been added, for removing columns in a
+  dataframe that present high correlation with other columns. :pr:`2023` by
+  :user:`Eloi Massoulié <emassoulie>`.
 
 Changes
 -------